Configs/mcp/.mcp/mcp_local/pyproject.toml
[project]
name = "local-mcp"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0.1",
  "psycopg[binary]>=3.2.0"
]

[tool.uv]
package = false
